+++ last_client_src/WidgetText.h
@@ -14,7 +14,7 @@ public:
 	WidgetText(void);
 	virtual ~WidgetText(void);
 
-	virtual void	loadSprites(std::list<std::pair<sf::Vector2f, sf::Vector2f> > &);
+	virtual void	loadSprites(std::list<std::pair<sf::Vector2i, sf::Vector2i> > &);
 	virtual void	setPosition(std::list<sf::Vector2f> &);
 	virtual void	setAnimations(std::list<std::pair<sf::Vector2f, float> > &);
 	virtual void	update(float, RenderWindow &, int);
