+++ last_client_src/WidgetTypingBar.cpp
@@ -6,7 +6,7 @@ WidgetTypingBar::WidgetTypingBar(void) : WidgetText()
 	this->setStyle(sf::String::Bold, 30, "Images/charlie_dotted.ttf", 255, 255, 255);
 	this->_text.SetText("| ");
 	this->_activated = false;
-	this->_timer = 0.6;
+	this->_timer = 0.6f;
 }
 
 
@@ -25,11 +25,11 @@ void		WidgetTypingBar::update(float elapsed, RenderWindow &win, int size)
 	}
 	if ((this->_timer -= elapsed) > 0.0)
 	{
-		this->_text.SetSize(size);
+		this->_text.SetSize(static_cast<float>(size));
 		win.drawText(this->_text);
 	}
-	else if (this->_timer <= -0.6)
-		this->_timer = 0.6;
+	else if (this->_timer <= -0.6f)
+		this->_timer = 0.6f;
 }
 
 void		WidgetTypingBar::switchActivate()
